requirements.txt
opencv-python==4.8.0.76
numpy==1.26.0
PyYAML==6.0.1
Pillow==10.1.0
matplotlib==3.8.1
matplotlib-inline==0.1.6
streamlit-drawable-canvas==0.9.3